+++ spotinst_sdk2/models/elastigroup/aws/asg.py
@@ -11,15 +11,38 @@ class ASG:
     product: str
     spot_instance_types: List[str]
     name: str
+    availability_vs_cost: str
     """
 
     def __init__(
             self,
             product=none,
             spot_instance_types=none,
-            name=none):
+            name=none,
+            availability_vs_cost=none):
 
         self.product = product
+        self.spot_instance_types = spot_instance_types
+        self.name = name
+        self.availability_vs_cost = availability_vs_cost
+
+# endregion
+
+# region Instance
+
+
+class ImportInstanceConfig:
+    """
+    # Arguments
+    spot_instance_types: List[str]
+    name: str
+    """
+
+    def __init__(
+            self,
+            spot_instance_types=none,
+            name=none):
+
         self.spot_instance_types = spot_instance_types
         self.name = name
 
@@ -27,6 +50,15 @@ class ASG:
 
 
 class ImportASGRequest:
+    def __init__(self, group):
+        self.group = group
+
+    def toJSON(self):
+        return json.dumps(self, default=lambda o: o.__dict__,
+                          sort_keys=True, indent=4)
+
+
+class ImportInstanceRequest:
     def __init__(self, group):
         self.group = group